updated main() function
It now shuffles the deck thoroughly
If you're wondering why the awkward interval of randomly generating a number, that's because it simulates a random break in the deck that the dealer makes when shuffling. The dealer tries to break the deck in half APPROXIMATEVELY, but not even him/herself can always split the deck perfectly in half. So I chose an interval such that the split is around half the deck. (15 - 37)
A dealer does this about two or three times from what I've seen irl.